% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{read_config}
\usage{
read_config(path)
}
\arguments{
\item{path}{The environment of project}
}
\value{

}
\description{
This function is called by the skeleton function and read the config
}
