% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(config, path)
}
\arguments{
\item{config}{List of configuration parameters}

\item{path}{The environment of project}
}
\value{

}
\description{
It's called by skeleton function and read the data
}
