% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_target}
\alias{read_target}
\title{read_target}
\usage{
read_target(config, path)
}
\arguments{
\item{config}{List of configuration parameters}

\item{path}{The environment of project}
}
\value{

}
\description{
It's called by skeleton function and read the target
}
\examples{

}
