% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ML.R
\name{to_ML}
\alias{to_ML}
\title{to_ML}
\usage{
to_ML(df_1, config)
}
\arguments{
\item{config}{List of configuration parameters}

\item{data}{DataFrame with reshape data from all datasets}
}
\description{
This function is called by skeleton and it hooks the df onto 
a machine learning model
}
